# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.AuthProvider:
    enum:
    - email
    - google
    - api_key
    type: string
    x-enum-varnames:
    - AuthProviderEmail
    - AuthProviderGoogle
    - AuthProviderAPIKey
...
  domain.User:
    properties:
      auth_provider:
        $ref: '#/definitions/domain.AuthProvider'
      created_at:
        type: string
      current_period_start:
        type: string
      documents_used_this_period:
        type: integer
      email:
        type: string
      email_verified:
        type: boolean
      email_verified_at:
        type: string
      full_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      monthly_document_limit:
        type: integer
      role:
        $ref: '#/definitions/domain.UserRole'
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
  domain.UserRole:
    enum:
    - admin
    - manager
    - member
    - viewer
    - free
    - service
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleManager
    - RoleMember
    - RoleViewer
    - RoleFree
    - RoleService
...
  handler.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  handler.CreateUserRequest:
    properties:
      email:
        example: jane.doe@acme.com
        type: string
      full_name:
        example: Jane Doe
        type: string
      password:
        example: securepassword123
        type: string
      role:
        allOf:
        - $ref: '#/definitions/domain.UserRole'
        example: member
    required:
    - email
    - password
    - role
    type: object
...
  handler.ErrorResponseBody:
    properties:
      error:
        $ref: '#/definitions/handler.APIError'
      success:
        example: false
        type: boolean
    type: object
...
  handler.MessageResponse:
    properties:
      message:
        example: operation completed successfully
        type: string
    type: object
  handler.PagMeta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  handler.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/handler.PagMeta'
      success:
        example: true
        type: boolean
    type: object
...
  handler.UpdateUserRequest:
    properties:
      email:
        example: jane.smith@acme.com
        type: string
      full_name:
        example: Jane Smith
        type: string
      is_active:
        example: true
        type: boolean
      role:
        allOf:
        - $ref: '#/definitions/domain.UserRole'
        example: manager
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@satvos.io
    name: API Support
  description: Multi-tenant document processing service with AI-powered invoice parsing
    and GST validation.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SATVOS API
  version: "1.0"
paths:
...
  /users:
    get:
      description: List all users in the tenant (admin only)
      parameters:
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - default: 20
        description: Limit for pagination (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.User'
                  type: array
                meta:
                  $ref: '#/definitions/handler.PagMeta'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "403":
          description: Forbidden - admin only
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user in the tenant (admin only)
      parameters:
      - description: User details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.User'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "403":
          description: Forbidden - admin only
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Create a user
      tags:
      - users
  /users/{id}:
    delete:
      description: Delete a user from the tenant (admin only)
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.MessageResponse'
              type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "403":
          description: Forbidden - admin only
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - users
    get:
      description: Get user details (self or admin access)
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.User'
              type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user details (self can update name/email, admin can update
        role/active status)
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.User'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and the JWT token. Example: Bearer
      eyJhbGci...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"